+++ src/main/java/nextstep/subway/line/application/LineService.java
@@ -42,6 +42,10 @@ public class LineService {
                 .collect(Collectors.toList());
     }
 
+    public List<Line> findAllLines() {
+        return lineRepository.findAll();
+    }
+
     public Line findLineById(Long id) {
         return lineRepository.findById(id).orElseThrow(() -> new NotExistException("해당 노선은 존재하지 않습니다."));
     }
