--- src/main/java/nextstep/subway/path/PathResponse.java
@@ -1,25 +1,0 @@
-package nextstep.subway.path;
-
-import java.util.List;
-import nextstep.subway.station.dto.StationResponse;
-
-public class PathResponse {
-    private List<StationResponse> stations;
-    private int distance;
-
-    public PathResponse() {
-    }
-
-    public PathResponse(List<StationResponse> stations, int distance) {
-        this.stations = stations;
-        this.distance = distance;
-    }
-
-    public List<StationResponse> getStations() {
-        return stations;
-    }
-
-    public int getDistance() {
-        return distance;
-    }
-}
