--- src/main/java/nextstep/subway/path/PathService.java
@@ -1,11 +1,0 @@
-package nextstep.subway.path;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class PathService {
-
-    public PathResponse searchShortestPath(Long source, Long target) {
-        return new PathResponse();
-    }
-}
