+++ src/main/java/nextstep/subway/path/application/PathService.java
@@ -1,0 +1,30 @@
+package nextstep.subway.path.application;
+
+import nextstep.subway.line.application.LineService;
+import nextstep.subway.path.domain.PathFinder;
+import nextstep.subway.path.dto.PathResponse;
+import nextstep.subway.station.application.StationService;
+import nextstep.subway.station.domain.Station;
+import org.springframework.stereotype.Service;
+
+@Service
+public class PathService {
+    private final StationService stationService;
+    private final LineService lineService;
+    private final PathFinder pathFinder;
+
+    public PathService(StationService stationService, LineService lineService,
+                       PathFinder pathFinder) {
+        this.stationService = stationService;
+        this.lineService = lineService;
+        this.pathFinder = pathFinder;
+    }
+
+    public PathResponse searchShortestPath(Long source, Long target) {
+        Station sourceStation = stationService.findStationById(source);
+        Station targetStation = stationService.findStationById(target);
+
+        pathFinder.init(lineService.findAllLines());
+        return PathResponse.of(pathFinder.getDijkstraPath(sourceStation, targetStation));
+    }
+}
