+++ src/main/java/nextstep/subway/path/ui/PathController.java
@@ -1,5 +1,7 @@
-package nextstep.subway.path;
+package nextstep.subway.path.ui;
 
+import nextstep.subway.path.dto.PathResponse;
+import nextstep.subway.path.application.PathService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
