+++ todo/step2.md
@@ -159,11 +159,11 @@ ex) 경로 조회를 수행하는 도메인 구현 예시
 ### 📚 Todo List 📚
 - [x] 리뷰 사항 반영
 - [ ] 최단 경로 조회 인수테스트 만들기
-  - [ ] 정상적인 출발, 도착역을 경로조회하면 최단거리를 알려준다.
+  - [x] 정상적인 출발, 도착역을 경로조회하면 최단거리를 알려준다.
   - [ ] 연결되어 있지 않은 출발, 도착역은 경로조회 할 수 없다.
   - [ ] 출발, 도착역이 동일하면 경로조회 할 수 없다.
   - [ ] 존재하지 않는 역은 경로조회 할 수 없다.
 - [ ] 최단 경로 조회 기능 구현하기
-  - [ ] PathController 만들기
+  - [x] PathController 만들기
   - [ ] PathService 만들기
-  - [ ] PathResponse 만들기
+  - [x] PathResponse 만들기
